Added ClassGraph#skipAllWriteOperations()
and NoopNestedJarHandler (to address #236)
